move drawer to separate file
