+++ LanguageApp/app/navigation/AppNavigator.js
@@ -10,19 +10,7 @@ import ProviderScenarioScreen from "../screens/ProviderScenarioScreen";
 import LearnerScenarioScreen from "../screens/LearnerScenarioScreen";
 import AboutScreen from "../screens/AboutScreen";
 import LoginScreen from "../screens/LoginScreen";
-
-import { createDrawerNavigator } from '@react-navigation/drawer';
-const Drawer = createDrawerNavigator();
-
-function MyDrawer(){
-  return(
-    <Drawer.Navigator>
-      <Drawer.Screen name="Choose User Type" component={UserTypeScreen} />
-      <Drawer.Screen name="Languages" component={LanguagesScreen} />
-      <Drawer.Screen name="Categories" component={CategoriesScreen} />
-    </Drawer.Navigator>
-  )
-}
+import MyDrawer from "./DrawerNavigator";
 
 const Stack = createStackNavigator();
 
@@ -34,7 +22,7 @@ const AppNavigator = () => (
       headerTitleStyle: { fontSize: 30, paddingBottom: 10 },
     }}
   >
-    <Stack.Screen name="I don't want this" component={MyDrawer} options={{headerShown:false}}/>
+    <Stack.Screen name="Menu" component={MyDrawer} options={{headerShown:false}}/>
     <Stack.Screen name="User Type" component={UserTypeScreen} />
     <Stack.Screen name="Languages" component={LanguagesScreen} />
     <Stack.Screen
